==> 0_mol_to_csv.ipynb <==
## Data Understanding
- load/observe mol files
- create csv files

import pandas as pd
import numpy as np
import re
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
warnings.simplefilter(action='ignore')

from glob import glob
from tqdm import tqdm

path = 'data/'

sample = pd.read_csv(path + 'sample_submission.csv')
train = pd.read_csv(path + 'train_set.ReorgE.csv')
test = pd.read_csv(path + 'test_set.csv')

train

## Loading MOL Files
- train_set
- test_set

### train_set

path = 'data/mol_files/train_set/'

train_mol = sorted(glob(path + '*.mol'))

train_mol[:10]

### test_set

path = 'data/mol_files/test_set/'

test_mol = sorted(glob(path + '*.mol'))

test_mol[:10]

## Observing MOL Files
- file structure
- length of an element

### File Structure

mol_file = open(train_mol[0], 'r').read().split("\n")
mol_file

##### Anatomy of a mol file #####

# Counts Block
# 53: number of atoms
# 55: number of bonds

# Atoms Block
# 3.0263   -2.3512    0.5507: x, y, z coordinates
# C: atom symbol

# Bonds Block(1  2  1  0)
# 1: first atom row number
# 2: second atom row number
# 1: bond type
# 0: bond stereochemistry

### Length of an Element
- atoms block
- bonds block

# atoms block
tmp = re.sub(' +', ' ', mol_file[4].lstrip()).split(' ')

print(tmp)
len(tmp)

# bonds block
tmp = re.sub(' +', ' ', mol_file[-4].lstrip()).split(' ')

print(tmp)
len(tmp)

length = []

for i in mol_file:
    tmp = re.sub(' +', ' ', i.lstrip()).split(' ')
    length.append(len(tmp))

set(length)

## Creating CSV Files
- atoms block
    - x, y, z coordinates
    - atom symbol
- bonds block
    - bond type

### train_set

# convert mol files to the form of dataframe(csv files)

for i in tqdm(train_mol):
    
    atoms = []
    bonds = []
    mol_file = open(i, 'r').read().split("\n")
    
    # create nested lists
    for j in mol_file:
        tmp = re.sub(' +', ' ', j.lstrip()).split(' ') # remove blanks
        
        if len(tmp) > 11: # atoms block
            atoms.append(tmp)
        elif 4 <= len(tmp) < 10: # bonds block
            bonds.append(tmp)
            
    file_name = i.split('\\')[-1].split('.')[0] # file name without filename extension(ex: train_0_g)
    path = 'data/mol_files/train_set/'
    
    # create csv files
    pd.DataFrame(atoms).to_csv(path + file_name + '_a.csv', index=False)
    pd.DataFrame(bonds).to_csv(path + file_name + '_b.csv', index=False)

path = 'data/mol_files/train_set/'

# atoms block
df = pd.read_csv(path + 'train_0_g_a.csv')
df.head()

# bonds block
df = pd.read_csv(path + 'train_0_g_b.csv')
df.head()

### test_set

for i in tqdm(test_mol):
    
    atoms = []
    bonds = []
    mol_file = open(i, 'r').read().split("\n")
    
    for j in mol_file:
        tmp = re.sub(' +', ' ', j.lstrip()).split(' ')
        
        if len(tmp) > 11:
            atoms.append(tmp)
        elif 4 <= len(tmp) < 10:
            bonds.append(tmp)
            
    file_name = i.split('\\')[-1].split('.')[0]
    path = 'data/mol_files/test_set/'
    
    pd.DataFrame(atoms).to_csv(path + file_name + '_a.csv', index=False)
    pd.DataFrame(bonds).to_csv(path + file_name + '_b.csv', index=False)
