==> 1-1_data_preparation_sample_demo.ipynb <==
## Data Preparation (1)
- create derived variables
    - Euclidean distance from the origin
    - atomic information from the periodic table
    - bond information
- sample demonstration

import pandas as pd
import numpy as np
import re
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
warnings.simplefilter(action='ignore')

from glob import glob
from tqdm import tqdm

path = 'data/'

sample = pd.read_csv(path + 'sample_submission.csv')
train = pd.read_csv(path + 'train_set.ReorgE.csv')
test = pd.read_csv(path + 'test_set.csv')

## Sample Demonstration on Creating Derived Variables
- Euclidean distance from the origin
- atomic information
    - numerical data from the periodic table
- bond information
    - bond length
    - bond energy

### (1) Distance from the Origin

path = 'data/mol_files/train_set/'

df_a = pd.read_csv(path + 'train_0_g_a.csv')
df_a.head()

df_a['dist'] = np.sqrt(df_a['0']**2 + df_a['1']**2 + df_a['2']**2)
df_a.head()

### (2) Atomic Information

# train_set

path = 'data/mol_files/train_set/'

symbols = []

for i in tqdm(train['index']):
    
    at_g = pd.read_csv(path + i + '_g_a.csv')
        
    for j in at_g['3']:
        symbols.append(j)

# test_set

path = 'data/mol_files/test_set/'

for i in tqdm(test['index']):
    
    at_g = pd.read_csv(path + i + '_g_a.csv')
        
    for j in at_g['3']:
        symbols.append(j)

symbols = list(set(symbols))
symbols

table = pd.read_csv('https://pubchem.ncbi.nlm.nih.gov/rest/pug/periodictable/CSV/?response_type=display')
table.head().T

table.index = table['Symbol']
tmp = table.loc[symbols].reset_index(drop=True)
tmp

##### numerical variables #####
# AtomicNumber, AtomicMass, Electronegativity, IonizationEnergy,
# ElectronAffinity, MeltingPoint, BoilingPoint, Density
# ** use 'bond length' instead of 'atomic radius'

##### categorical variables #####
# StandardState, GroupBlock

at_number = {}

for i in range(12):
    key = tmp.iloc[i]['Symbol']
    val = tmp.iloc[i]['AtomicNumber']
    
    at_number[key] = val

at_number

at_weight = {}

for i in range(12):
    key = tmp.iloc[i]['Symbol']
    val = tmp.iloc[i]['AtomicMass']
    
    at_weight[key] = val

electro_neg = {}

for i in range(12):
    key = tmp.iloc[i]['Symbol']
    val = tmp.iloc[i]['Electronegativity']
    
    electro_neg[key] = val

ion_energy = {}

for i in range(12):
    key = tmp.iloc[i]['Symbol']
    val = tmp.iloc[i]['IonizationEnergy']
    
    ion_energy[key] = val

electron_aff = {}

for i in range(12):
    key = tmp.iloc[i]['Symbol']
    val = tmp.iloc[i]['ElectronAffinity']
    
    electron_aff[key] = val

melting_p = {}

for i in range(12):
    key = tmp.iloc[i]['Symbol']
    val = tmp.iloc[i]['MeltingPoint']
    
    melting_p[key] = val

boiling_p = {}

for i in range(12):
    key = tmp.iloc[i]['Symbol']
    val = tmp.iloc[i]['BoilingPoint']
    
    boiling_p[key] = val

density = {}

for i in range(12):
    key = tmp.iloc[i]['Symbol']
    val = tmp.iloc[i]['Density']
    
    density[key] = val

# add atomic information to df_a as variables
df_a['at_n'] = df_a['3'].map(at_number)
df_a['at_w'] = df_a['3'].map(at_weight)
df_a['el_neg'] = df_a['3'].map(electro_neg)
df_a['ion_en'] = df_a['3'].map(ion_energy)
df_a['el_aff'] = df_a['3'].map(electron_aff)
df_a['mp'] = df_a['3'].map(melting_p)
df_a['bp'] = df_a['3'].map(boiling_p)
df_a['den'] = df_a['3'].map(density)
df_a.head()

### (3) Bond Information
- bond length
- bond energy

path = 'data/mol_files/train_set/'

df_b = pd.read_csv(path + 'train_0_g_b.csv')
df_b.head()

df_a.index = range(1, len(df_a)+1)

df_a.head()

#### bond length

at_list_1 = []
at_list_2 = []
len_list = []

for i in df_b.index:
    
    idx_1 = df_b.iloc[i, 0]
    idx_2 = df_b.iloc[i, 1]
    
    atom_1 = df_a.loc[idx_1]
    atom_2 = df_a.loc[idx_2]
    
    sym_1 = atom_1['3']
    sym_2 = atom_2['3']
    bond_len = np.sqrt((atom_1['0'] - atom_2['0'])**2 + (atom_1['1'] - atom_2['1'])**2 + (atom_1['2'] - atom_2['2'])**2)
    
    at_list_1.append(sym_1)
    at_list_2.append(sym_2)
    len_list.append(bond_len)

bond_df = pd.DataFrame()
bond_df['at_1'] = at_list_1
bond_df['at_2'] = at_list_2
bond_df['type'] = df_b['2']
bond_df['len'] = len_list

bond_df.head()

# bond_df contains bond information such as bond length and bond energy
# interpreted from bonds block data in df_b
# thus, len(bond_df) == len(df_b)

#### bond energy

table = pd.read_html('https://www.wiredchemist.com/chemistry/data/bond_energies_lengths.html')

t1 = table[0]
t2 = table[1]
t3 = table[2]
t4 = table[3]
t5 = table[4].loc[0:22]
t6 = table[5].drop(4)
t7 = table[6]

t12 = pd.concat([t1, t2])
t34 = pd.concat([t3, t4])
t56 = pd.concat([t5, t6])

t1234 = pd.concat([t12, t34])
t567 = pd.concat([t56, t7])
t_all = pd.concat([t1234, t567])

en_df = t_all.iloc[:, 0:2].dropna().reset_index(drop=True)
en_df.columns = ['bond', 'energy']
en_df

en_df['type'] = 1
en_df['at_1'] = 1
en_df['at_2'] = 1
en_df.head()

for idx, bond in enumerate(en_df['bond']):
    if '-' in bond:
        at_1 = bond.split('-')[0]
        at_2 = bond.split('-')[1]
        en_df.loc[idx, 'at_1'] = at_1
        en_df.loc[idx, 'at_2'] = at_2
    elif '=' in bond:
        at_1 = bond.split('=')[0]
        at_2 = bond.split('=')[1]
        en_df.loc[idx, 'type'] = 2
        en_df.loc[idx, 'at_1'] = at_1
        en_df.loc[idx, 'at_2'] = at_2
    elif '≡' in bond:
        at_1 = bond.split('≡')[0]
        at_2 = bond.split('≡')[1]
        en_df.loc[idx, 'type'] = 3
        en_df.loc[idx, 'at_1'] = at_1
        en_df.loc[idx, 'at_2'] = at_2

en_df.head()

bond_df['bond_en'] = 1
bond_df.head()

set_list = []

for i in bond_df.index:
    at_set = {bond_df.loc[i, 'at_1'], bond_df.loc[i, 'at_2']}
    set_list.append(at_set)

set_list[:10]

for idx, bond in enumerate(set_list):
    
    for j in en_df.index:
        at_set = {en_df.loc[j, 'at_1'], en_df.loc[j, 'at_2']}
        
        if (bond == at_set) and (bond_df.loc[idx, 'type'] == en_df.loc[j, 'type']):
            bond_df.loc[idx, 'bond_en'] = en_df.loc[j, 'energy']

bond_df.head()

## Aggregation of Created Variables
- np.mean

df_a.head()

bond_df.head()

df_a.columns

bond_df.columns

a_cols = ['at_n', 'at_w', 'el_neg', 'ion_en', 'el_aff', 'mp', 'bp', 'den']
b_cols = ['len', 'bond_en']

# use np.mean as agg function since none of the atoms in a molecule are considered an outlier

data = []
avg_dic = {}

for i in a_cols:
    avg = np.mean(df_a[i])
    avg_dic[i] = avg

for i in b_cols:
    avg = np.mean(bond_df[i])
    avg_dic[i] = avg

data.append(avg_dic)

pd.DataFrame(data)
