==> 1-2_data_preparation_complete_data.ipynb <==
## Data Preparation (2)
- derive variables from complete data

import pandas as pd
import numpy as np
import re
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
warnings.simplefilter(action='ignore')

from glob import glob
from tqdm import tqdm

path = 'data/'

sample = pd.read_csv(path + 'sample_submission.csv')
train = pd.read_csv(path + 'train_set.ReorgE.csv')
test = pd.read_csv(path + 'test_set.csv')

train

## Deriving Variables from Complete Data
- create derived variables
    - Euclidean distance from the origin
    - atomic information from the periodic table
    - bond information
- aggregation
    - np.mean

### Preparing Materials
- symbols list
- dictionaries for mapping
- bond energy table

#### symbols list

symbols = ['N', 'C', 'Si', 'I', 'B', 'O', 'P', 'Br', 'S', 'F', 'H', 'Cl']

#### dictionaries for mapping

table = pd.read_csv('https://pubchem.ncbi.nlm.nih.gov/rest/pug/periodictable/CSV/?response_type=display')
table.index = table['Symbol']
tmp = table.loc[symbols].reset_index(drop=True)

tmp.columns

at_number = {}
at_weight = {}
electro_neg = {}
ion_energy = {}
electron_aff = {}
melting_p = {}
boiling_p = {}
density = {}

for i in range(12):
    key = tmp.iloc[i]['Symbol']
    val1 = tmp.iloc[i]['AtomicNumber']
    val2 = tmp.iloc[i]['AtomicMass']
    val3 = tmp.iloc[i]['Electronegativity']
    val4 = tmp.iloc[i]['IonizationEnergy']
    val5 = tmp.iloc[i]['ElectronAffinity']
    val6 = tmp.iloc[i]['MeltingPoint']
    val7 = tmp.iloc[i]['BoilingPoint']
    val8 = tmp.iloc[i]['Density']
    
    at_number[key] = val1
    at_weight[key] = val2
    electro_neg[key] = val3
    ion_energy[key] = val4
    electron_aff[key] = val5
    melting_p[key] = val6
    boiling_p[key] = val7
    density[key] = val8

#### bond energy table

table = pd.read_html('https://www.wiredchemist.com/chemistry/data/bond_energies_lengths.html')

t1 = table[0]
t2 = table[1]
t3 = table[2]
t4 = table[3]
t5 = table[4].loc[0:22]
t6 = table[5].drop(4)
t7 = table[6]

t12 = pd.concat([t1, t2])
t34 = pd.concat([t3, t4])
t56 = pd.concat([t5, t6])

t1234 = pd.concat([t12, t34])
t567 = pd.concat([t56, t7])
t_all = pd.concat([t1234, t567])

en_df = t_all.iloc[:, 0:2].dropna().reset_index(drop=True)
en_df.columns = ['bond', 'energy']

en_df['type'] = 1
en_df['at_1'] = 1
en_df['at_2'] = 1

for idx, bond in enumerate(en_df['bond']):
    if '-' in bond:
        at_1 = bond.split('-')[0]
        at_2 = bond.split('-')[1]
        en_df.loc[idx, 'at_1'] = at_1
        en_df.loc[idx, 'at_2'] = at_2
    elif '=' in bond:
        at_1 = bond.split('=')[0]
        at_2 = bond.split('=')[1]
        en_df.loc[idx, 'type'] = 2
        en_df.loc[idx, 'at_1'] = at_1
        en_df.loc[idx, 'at_2'] = at_2
    elif '≡' in bond:
        at_1 = bond.split('≡')[0]
        at_2 = bond.split('≡')[1]
        en_df.loc[idx, 'type'] = 3
        en_df.loc[idx, 'at_1'] = at_1
        en_df.loc[idx, 'at_2'] = at_2

### Creating New Dataset
- train_set
    - train_g, train_ex
- test_set
    - test_g, test_ex

#### train_set

##### ground state #####

data_g = []
path = 'data/mol_files/train_set/'

for name in tqdm(train['index']):
    
    g_a = pd.read_csv(path + name + '_g_a.csv')
    g_b = pd.read_csv(path + name + '_g_b.csv')
    
    
    # creating derived variables
    
    # (1) distance from the origin
    g_a['dist'] = np.sqrt(g_a['0']**2 + g_a['1']**2 + g_a['2']**2)
    
    # (2) atomic information
    g_a['at_n'] = g_a['3'].map(at_number)
    g_a['at_w'] = g_a['3'].map(at_weight)
    g_a['el_neg'] = g_a['3'].map(electro_neg)
    g_a['ion_en'] = g_a['3'].map(ion_energy)
    g_a['el_aff'] = g_a['3'].map(electron_aff)
    g_a['mp'] = g_a['3'].map(melting_p)
    g_a['bp'] = g_a['3'].map(boiling_p)
    g_a['den'] = g_a['3'].map(density)
    
    # (3) bond information
    g_a.index = range(1, len(g_a)+1)
    
    # bond length
    at_list_1 = []
    at_list_2 = []
    len_list = []

    for i in g_b.index:

        idx_1 = g_b.iloc[i, 0]
        idx_2 = g_b.iloc[i, 1]

        atom_1 = g_a.loc[idx_1]
        atom_2 = g_a.loc[idx_2]

        sym_1 = atom_1['3']
        sym_2 = atom_2['3']
        bond_len = np.sqrt((atom_1['0'] - atom_2['0'])**2 + (atom_1['1'] - atom_2['1'])**2 + (atom_1['2'] - atom_2['2'])**2)

        at_list_1.append(sym_1)
        at_list_2.append(sym_2)
        len_list.append(bond_len)
    
    bond_df = pd.DataFrame()
    bond_df['at_1'] = at_list_1
    bond_df['at_2'] = at_list_2
    bond_df['type'] = g_b['2']
    bond_df['len'] = len_list
    
    # bond energy
    bond_df['bond_en'] = 1
    set_list = []

    for i in bond_df.index:
        at_set = {bond_df.loc[i, 'at_1'], bond_df.loc[i, 'at_2']}
        set_list.append(at_set)

    for idx, bond in enumerate(set_list):

        for j in en_df.index:
            at_set = {en_df.loc[j, 'at_1'], en_df.loc[j, 'at_2']}

            if (bond == at_set) and (bond_df.loc[idx, 'type'] == en_df.loc[j, 'type']):
                bond_df.loc[idx, 'bond_en'] = en_df.loc[j, 'energy']
    
    
    # aggregation of created variables
    
    a_cols = ['dist', 'at_n', 'at_w', 'el_neg', 'ion_en', 'el_aff', 'mp', 'bp', 'den']
    b_cols = ['len', 'bond_en']
    
    avg_dic = {}

    for i in a_cols:
        avg = np.mean(g_a[i])
        avg_dic[i] = avg

    for j in b_cols:
        avg = np.mean(bond_df[j])
        avg_dic[j] = avg

    data_g.append(avg_dic)

df_g = pd.DataFrame(data_g)
df_g

##### exited state #####

data_ex = []
path = 'data/mol_files/train_set/'

for name in tqdm(train['index']):
    
    ex_a = pd.read_csv(path + name + '_ex_a.csv')
    ex_b = pd.read_csv(path + name + '_ex_b.csv')
    
    
    # creating derived variables
    
    # (1) distance from the origin
    ex_a['dist'] = np.sqrt(ex_a['0']**2 + ex_a['1']**2 + ex_a['2']**2)
    
    # (2) atomic information
    ex_a['at_n'] = ex_a['3'].map(at_number)
    ex_a['at_w'] = ex_a['3'].map(at_weight)
    ex_a['el_neg'] = ex_a['3'].map(electro_neg)
    ex_a['ion_en'] = ex_a['3'].map(ion_energy)
    ex_a['el_aff'] = ex_a['3'].map(electron_aff)
    ex_a['mp'] = ex_a['3'].map(melting_p)
    ex_a['bp'] = ex_a['3'].map(boiling_p)
    ex_a['den'] = ex_a['3'].map(density)
    
    # (3) bond information
    ex_a.index = range(1, len(ex_a)+1)
    
    # bond length
    at_list_1 = []
    at_list_2 = []
    len_list = []

    for i in ex_b.index:

        idx_1 = ex_b.iloc[i, 0]
        idx_2 = ex_b.iloc[i, 1]

        atom_1 = ex_a.loc[idx_1]
        atom_2 = ex_a.loc[idx_2]

        sym_1 = atom_1['3']
        sym_2 = atom_2['3']
        bond_len = np.sqrt((atom_1['0'] - atom_2['0'])**2 + (atom_1['1'] - atom_2['1'])**2 + (atom_1['2'] - atom_2['2'])**2)

        at_list_1.append(sym_1)
        at_list_2.append(sym_2)
        len_list.append(bond_len)
    
    bond_df = pd.DataFrame()
    bond_df['at_1'] = at_list_1
    bond_df['at_2'] = at_list_2
    bond_df['type'] = ex_b['2']
    bond_df['len'] = len_list
    
    # bond energy
    bond_df['bond_en'] = 1
    set_list = []

    for i in bond_df.index:
        at_set = {bond_df.loc[i, 'at_1'], bond_df.loc[i, 'at_2']}
        set_list.append(at_set)

    for idx, bond in enumerate(set_list):

        for j in en_df.index:
            at_set = {en_df.loc[j, 'at_1'], en_df.loc[j, 'at_2']}

            if (bond == at_set) and (bond_df.loc[idx, 'type'] == en_df.loc[j, 'type']):
                bond_df.loc[idx, 'bond_en'] = en_df.loc[j, 'energy']
    
    
    # aggregation of created variables
    
    a_cols = ['dist', 'at_n', 'at_w', 'el_neg', 'ion_en', 'el_aff', 'mp', 'bp', 'den']
    b_cols = ['len', 'bond_en']
    
    avg_dic = {}

    for i in a_cols:
        avg = np.mean(ex_a[i])
        avg_dic[i] = avg

    for j in b_cols:
        avg = np.mean(bond_df[j])
        avg_dic[j] = avg

    data_ex.append(avg_dic)

df_ex = pd.DataFrame(data_ex)
df_ex

df_g.to_csv('train_g.csv', index=False)
df_ex.to_csv('train_ex.csv', index=False)

#### test_set

##### ground state #####

data_g = []
path = 'data/mol_files/test_set/'

for name in tqdm(test['index']):
    
    g_a = pd.read_csv(path + name + '_g_a.csv')
    g_b = pd.read_csv(path + name + '_g_b.csv')
    
    
    # creating derived variables
    
    # (1) distance from the origin
    g_a['dist'] = np.sqrt(g_a['0']**2 + g_a['1']**2 + g_a['2']**2)
    
    # (2) atomic information
    g_a['at_n'] = g_a['3'].map(at_number)
    g_a['at_w'] = g_a['3'].map(at_weight)
    g_a['el_neg'] = g_a['3'].map(electro_neg)
    g_a['ion_en'] = g_a['3'].map(ion_energy)
    g_a['el_aff'] = g_a['3'].map(electron_aff)
    g_a['mp'] = g_a['3'].map(melting_p)
    g_a['bp'] = g_a['3'].map(boiling_p)
    g_a['den'] = g_a['3'].map(density)
    
    # (3) bond information
    g_a.index = range(1, len(g_a)+1)
    
    # bond length
    at_list_1 = []
    at_list_2 = []
    len_list = []

    for i in g_b.index:

        idx_1 = g_b.iloc[i, 0]
        idx_2 = g_b.iloc[i, 1]

        atom_1 = g_a.loc[idx_1]
        atom_2 = g_a.loc[idx_2]

        sym_1 = atom_1['3']
        sym_2 = atom_2['3']
        bond_len = np.sqrt((atom_1['0'] - atom_2['0'])**2 + (atom_1['1'] - atom_2['1'])**2 + (atom_1['2'] - atom_2['2'])**2)

        at_list_1.append(sym_1)
        at_list_2.append(sym_2)
        len_list.append(bond_len)
    
    bond_df = pd.DataFrame()
    bond_df['at_1'] = at_list_1
    bond_df['at_2'] = at_list_2
    bond_df['type'] = g_b['2']
    bond_df['len'] = len_list
    
    # bond energy
    bond_df['bond_en'] = 1
    set_list = []

    for i in bond_df.index:
        at_set = {bond_df.loc[i, 'at_1'], bond_df.loc[i, 'at_2']}
        set_list.append(at_set)

    for idx, bond in enumerate(set_list):

        for j in en_df.index:
            at_set = {en_df.loc[j, 'at_1'], en_df.loc[j, 'at_2']}

            if (bond == at_set) and (bond_df.loc[idx, 'type'] == en_df.loc[j, 'type']):
                bond_df.loc[idx, 'bond_en'] = en_df.loc[j, 'energy']
    
    
    # aggregation of created variables
    
    a_cols = ['dist', 'at_n', 'at_w', 'el_neg', 'ion_en', 'el_aff', 'mp', 'bp', 'den']
    b_cols = ['len', 'bond_en']
    
    avg_dic = {}

    for i in a_cols:
        avg = np.mean(g_a[i])
        avg_dic[i] = avg

    for j in b_cols:
        avg = np.mean(bond_df[j])
        avg_dic[j] = avg

    data_g.append(avg_dic)

df_g = pd.DataFrame(data_g)

##### exited state #####

data_ex = []
path = 'data/mol_files/test_set/'

for name in tqdm(test['index']):
    
    ex_a = pd.read_csv(path + name + '_ex_a.csv')
    ex_b = pd.read_csv(path + name + '_ex_b.csv')
    
    
    # creating derived variables
    
    # (1) distance from the origin
    ex_a['dist'] = np.sqrt(ex_a['0']**2 + ex_a['1']**2 + ex_a['2']**2)
    
    # (2) atomic information
    ex_a['at_n'] = ex_a['3'].map(at_number)
    ex_a['at_w'] = ex_a['3'].map(at_weight)
    ex_a['el_neg'] = ex_a['3'].map(electro_neg)
    ex_a['ion_en'] = ex_a['3'].map(ion_energy)
    ex_a['el_aff'] = ex_a['3'].map(electron_aff)
    ex_a['mp'] = ex_a['3'].map(melting_p)
    ex_a['bp'] = ex_a['3'].map(boiling_p)
    ex_a['den'] = ex_a['3'].map(density)
    
    # (3) bond information
    ex_a.index = range(1, len(ex_a)+1)
    
    # bond length
    at_list_1 = []
    at_list_2 = []
    len_list = []

    for i in ex_b.index:

        idx_1 = ex_b.iloc[i, 0]
        idx_2 = ex_b.iloc[i, 1]

        atom_1 = ex_a.loc[idx_1]
        atom_2 = ex_a.loc[idx_2]

        sym_1 = atom_1['3']
        sym_2 = atom_2['3']
        bond_len = np.sqrt((atom_1['0'] - atom_2['0'])**2 + (atom_1['1'] - atom_2['1'])**2 + (atom_1['2'] - atom_2['2'])**2)

        at_list_1.append(sym_1)
        at_list_2.append(sym_2)
        len_list.append(bond_len)
    
    bond_df = pd.DataFrame()
    bond_df['at_1'] = at_list_1
    bond_df['at_2'] = at_list_2
    bond_df['type'] = ex_b['2']
    bond_df['len'] = len_list
    
    # bond energy
    bond_df['bond_en'] = 1
    set_list = []

    for i in bond_df.index:
        at_set = {bond_df.loc[i, 'at_1'], bond_df.loc[i, 'at_2']}
        set_list.append(at_set)

    for idx, bond in enumerate(set_list):

        for j in en_df.index:
            at_set = {en_df.loc[j, 'at_1'], en_df.loc[j, 'at_2']}

            if (bond == at_set) and (bond_df.loc[idx, 'type'] == en_df.loc[j, 'type']):
                bond_df.loc[idx, 'bond_en'] = en_df.loc[j, 'energy']
    
    
    # aggregation of created variables
    
    a_cols = ['dist', 'at_n', 'at_w', 'el_neg', 'ion_en', 'el_aff', 'mp', 'bp', 'den']
    b_cols = ['len', 'bond_en']
    
    avg_dic = {}

    for i in a_cols:
        avg = np.mean(ex_a[i])
        avg_dic[i] = avg

    for j in b_cols:
        avg = np.mean(bond_df[j])
        avg_dic[j] = avg

    data_ex.append(avg_dic)

df_ex = pd.DataFrame(data_ex)

df_g.to_csv('test_g.csv', index=False)
df_ex.to_csv('test_ex.csv', index=False)
