==> 2_modeling.ipynb <==
## Modeling
- data preparation for modeling
- design machine learning model

import pandas as pd
import numpy as np
import re
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
warnings.simplefilter(action='ignore')

from glob import glob
from tqdm import tqdm

path = 'data/'

sample = pd.read_csv(path + 'sample_submission.csv')
train = pd.read_csv(path + 'train_set.ReorgE.csv')
test = pd.read_csv(path + 'test_set.csv')

train

### Preparing Data for Modeling
- reorganize train dataset
    - define feature and label
    - train test split

train_g = pd.read_csv('train_g.csv')
train_ex = pd.read_csv('train_ex.csv')

train_g['Reorg'] = train['Reorg_g']
train_ex['Reorg'] = train['Reorg_ex']

train_g.head()

train_ex.head()

train_set = pd.concat([train_g, train_ex], ignore_index=True)
train_set

train_set.to_csv('train_set.csv', index=False)

#### define feature and label

train_set = pd.read_csv('train_set.csv')

train_set.columns

features = ['dist', 'at_n', 'at_w', 'el_neg', 'ion_en',
            'el_aff', 'mp', 'bp', 'den', 'len', 'bond_en']
label = 'Reorg'

#### train test split
- systematic sampling

train_df = train_set[0::2].reset_index()
test_df = train_set[1::2].reset_index()

X_train, y_train = train_df[features], train_df[label]
X_test, y_test = test_df[features], test_df[label]

### Modeling and Evaluation

from sklearn.ensemble import RandomForestRegressor as rf

model = rf(n_jobs=-1)
model.fit(X_train, y_train)

model.score(X_train, y_train), model.score(X_test, y_test)

#### log transformation

train_set['Reorg'] = np.log(train_set['Reorg'])

train_df = train_set[0::2].reset_index()
test_df = train_set[1::2].reset_index()

X_train, y_train = train_df[features], train_df[label]
X_test, y_test = test_df[features], test_df[label]

log_m = rf(n_jobs=-1)
log_m.fit(X_train, y_train)

log_m.score(X_train, y_train), log_m.score(X_test, y_test)

## Submission
- prepare test_set
- modeling and evaluation

### Preparing test_set

test_g = pd.read_csv('test_g.csv')
test_ex = pd.read_csv('test_ex.csv')

test_set = pd.concat([test_g, test_ex], ignore_index=True)

test_set.to_csv('test_set.csv', index=False)

### Modeling and Evaluation

train_set = pd.read_csv('train_set.csv')
test_set = pd.read_csv('test_set.csv')

train_set.head()

test_set.head()

X_train, y_train = train_set[features], train_set[label]
X_test = test_set

model = rf(n_jobs=-1)
model.fit(X_train, y_train)

pred = model.predict(X_test)

sample['Reorg_g'] = pred[:457]
sample['Reorg_ex'] = pred[457:]

sample.to_csv('submission_2.csv', index=False)

#### log transformed model

sample = pd.read_csv(path + 'sample_submission.csv')
train_set = pd.read_csv('train_set.csv')
test_set = pd.read_csv('test_set.csv')

X_train, y_train = train_set[features], np.log1p(train_set[label])
X_test = test_set

model = rf(n_jobs=-1)
model.fit(X_train, y_train)

pred = np.expm1(model.predict(X_test))

sample['Reorg_g'] = pred[:457]
sample['Reorg_ex'] = pred[457:]

sample.to_csv('submission_3.csv', index=False)
